refactor(codebase): Remove langdetect package
The "LanguageDetection" folder as well as all its sub-folders has been entirely
removed from the codebase. Removing such a module /langdetect adds more
specificity as it belongs to an external project.
This reformulation does not affect the rest of the codebase
in the /ResumeParsing project.

Please refer to the appropriate project repo $, for more infos on the module.

Co-Authored-By: akram benmira <91384624+akrambenmira@users.noreply.github.com>