refactor(codebase): Remove the ResumeParsing repository
This commit intend to remove the entire /ResumeParsing repo from the codebase.
This reformulation is due to a radical organization of the java codebase, and thus,
the codebase will be affected by this changes as it heavily depends on its
sub-folders (i.e. /resources) and files (e.g. .class, .xml, .java, …etc.).

BREAKING CHANGE: It is noted that the current folder will be added as soon as possible,
and the codebase would be functionable by the next upcoming commits.